improve decodeColumn
