+++ core/src/main/scala/com/pingcap/tispark/utils/TiUtil.scala
@@ -208,6 +208,7 @@ object TiUtil {
     if (conf.contains(TiConfigConst.ENABLE_ARROW)) {
       tiConf.setEnableArrow(conf.get(TiConfigConst.ENABLE_ARROW).toBoolean)
     }
+
     if (conf.contains(TiConfigConst.REGION_INDEX_SCAN_DOWNGRADE_THRESHOLD)) {
       tiConf.setDowngradeThreshold(
         conf.get(TiConfigConst.REGION_INDEX_SCAN_DOWNGRADE_THRESHOLD).toInt
