+++ tikv-client/src/main/java/com/pingcap/tikv/columnar/TiColumnarBatch.java
@@ -1,7 +1,5 @@
 package com.pingcap.tikv.columnar;
 
-import org.apache.spark.sql.vectorized.ColumnarBatch;
-
 /**
  * This class wraps multiple ColumnVectors as a row-wise table. It provides a row view of this batch
  * so that Spark can access the data row by row. Instance of it is meant to be reused during the
