+++ tikv-client/src/main/java/com/pingcap/tikv/types/BytesType.java
@@ -22,7 +22,6 @@ import com.pingcap.tikv.codec.Codec;
 import com.pingcap.tikv.codec.Codec.BytesCodec;
 import com.pingcap.tikv.codec.CodecDataInput;
 import com.pingcap.tikv.codec.CodecDataOutput;
-import com.pingcap.tikv.columnar.TiChunkColumn;
 import com.pingcap.tikv.exception.ConvertNotSupportException;
 import com.pingcap.tikv.exception.ConvertOverflowException;
 import com.pingcap.tikv.exception.InvalidCodecFormatException;
@@ -73,11 +72,6 @@ public class BytesType extends DataType {
     } else {
       throw new InvalidCodecFormatException("Invalid Flag type for : " + flag);
     }
-  }
-
-  @Override
-  public TiChunkColumn decodeColumn(CodecDataInput cdi) {
-    return null;
   }
 
   @Override
