+++ tikv-client/src/main/java/com/pingcap/tikv/types/DateTimeType.java
@@ -19,7 +19,6 @@ package com.pingcap.tikv.types;
 
 import com.pingcap.tikv.ExtendedDateTime;
 import com.pingcap.tikv.codec.CodecDataInput;
-import com.pingcap.tikv.columnar.TiChunkColumn;
 import com.pingcap.tikv.exception.ConvertNotSupportException;
 import com.pingcap.tikv.exception.ConvertOverflowException;
 import com.pingcap.tikv.meta.TiColumnInfo;
@@ -66,11 +65,6 @@ public class DateTimeType extends AbstractDateTimeType {
       return null;
     }
     return extendedDateTime.toTimeStamp().getTime() * 1000;
-  }
-
-  @Override
-  public TiChunkColumn decodeColumn(CodecDataInput cdi) {
-    return null;
   }
 
   @Override
