+++ tikv-client/src/main/java/com/pingcap/tikv/types/DecimalType.java
@@ -23,7 +23,6 @@ import com.pingcap.tikv.codec.Codec.DecimalCodec;
 import com.pingcap.tikv.codec.CodecDataInput;
 import com.pingcap.tikv.codec.CodecDataOutput;
 import com.pingcap.tikv.codec.MyDecimal;
-import com.pingcap.tikv.columnar.TiChunkColumn;
 import com.pingcap.tikv.exception.ConvertNotSupportException;
 import com.pingcap.tikv.exception.ConvertOverflowException;
 import com.pingcap.tikv.exception.InvalidCodecFormatException;
@@ -54,11 +53,6 @@ public class DecimalType extends DataType {
       throw new InvalidCodecFormatException("Invalid Flag type for decimal type: " + flag);
     }
     return DecimalCodec.readDecimal(cdi);
-  }
-
-  @Override
-  public TiChunkColumn decodeColumn(CodecDataInput cdi) {
-    return null;
   }
 
   @Override
