+++ tikv-client/src/main/java/com/pingcap/tikv/types/IntegerType.java
@@ -23,7 +23,6 @@ import com.pingcap.tikv.codec.Codec;
 import com.pingcap.tikv.codec.Codec.IntegerCodec;
 import com.pingcap.tikv.codec.CodecDataInput;
 import com.pingcap.tikv.codec.CodecDataOutput;
-import com.pingcap.tikv.columnar.TiChunkColumn;
 import com.pingcap.tikv.exception.ConvertNotSupportException;
 import com.pingcap.tikv.exception.ConvertOverflowException;
 import com.pingcap.tikv.exception.TypeException;
@@ -108,25 +107,6 @@ public class IntegerType extends DataType {
       return unsignedValueOf(ret);
     }
     return ret;
-  }
-
-  @Override
-  public TiChunkColumn decodeColumn(CodecDataInput cdi) {
-    int numRows = 0;
-    int numNulls = 0;
-    int numNullBitmapBytes = (numRows + 7) / 8;
-    // first int is col's length
-    // second int is col's null count
-    // if null count is not 0
-    // int numNullBitmapBytes = (col.length + 7) / 8
-    // buffer(0, numNullBitmapBytes) ->
-    // if null count is 0, set all data is not null
-    // int numFixedBytes = getFixLen
-    // int numDataBytes =  numFixedBytes * col.length
-    // if elebuf's length is smaller than  numFixedBytes
-    // copy buffer[numNullBitmapBytes,numDataBytes] to column
-    // TODO make it work
-    return new TiChunkColumn(this, numRows, numNulls, null, null);
   }
 
   /** {@inheritDoc} */
