+++ tikv-client/src/main/java/com/pingcap/tikv/types/TimeType.java
@@ -22,7 +22,6 @@ import com.pingcap.tikv.codec.Codec;
 import com.pingcap.tikv.codec.Codec.IntegerCodec;
 import com.pingcap.tikv.codec.CodecDataInput;
 import com.pingcap.tikv.codec.CodecDataOutput;
-import com.pingcap.tikv.columnar.TiChunkColumn;
 import com.pingcap.tikv.exception.ConvertNotSupportException;
 import com.pingcap.tikv.exception.ConvertOverflowException;
 import com.pingcap.tikv.exception.TypeException;
@@ -64,11 +63,6 @@ public class TimeType extends DataType {
       return IntegerCodec.readLong(cdi);
     }
     throw new TypeException("Invalid TimeType flag: " + flag);
-  }
-
-  @Override
-  public TiChunkColumn decodeColumn(CodecDataInput cdi) {
-    return null;
   }
 
   @Override
